Fix: 유저 chatId 로직 userService에서 분리, 빈 등록
